requirements.txt
gradio
gradio_imageslider
torch
torchvision
pillow
numpy
typing
gitpython
huggingface_hub
opencv-python
scikit-image
moviepy